backend/requirements.txt
groq
fastapi
uvicorn[standard]
xgboost
shap
scikit-learn
pandas
numpy
joblib
pydantic
python-multipart
matplotlib
fairlearn

pymongo==4.6.0
python-dotenv==1.0.0
